requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
channels==3.0.5
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.4
daphne==3.0.2
Django==3.2
djangorestframework==3.14.0
gevent==23.9.1
greenlet==2.0.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pytz==2023.3.post1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.0.5
websocket==0.2.1
websocket-client==1.6.3
websockets==11.0.3
zope.event==5.0
zope.interface==6.0
